windows test only + moved to correct location
